+++ backend/test/integration/ogpControll.test.ts
@@ -2,6 +2,9 @@ import { describe, it, expect } from 'vitest';
 import getMeta from '@/api/v1.6/ogp/meta.get.service.js';
 
 describe('getOgp Test', () => {
+    const launchEnv = process.env.LAUNCH_ENV || 'local';
+    const isLocal = launchEnv === 'local';
+
     // YoutubeはCORSが設定されている例
     it('True Website cors site test', async () => {
         const result = await getMeta({
@@ -68,34 +71,34 @@ describe('getOgp Test', () => {
     }, 50000);
 
     //twitter.com/nekono_dev/status/1774369918726947311
-    it('Bad request localhost test', async () => {
+    (isLocal ? it.skip : it)('Bad request localhost test', async () => {
         const result = await getMeta({
             url: 'http://localhost/',
             lang: 'ja',
         });
         expect(result).toEqual({
             success: false,
-            error: 'InternalServerError',
+            error: 'BadRequest',
         });
     }, 50000);
-    it('Bad request no dns ipv4 website test', async () => {
+    (isLocal ? it.skip : it)('Bad request no dns ipv4 website test', async () => {
         const result = await getMeta({
             url: 'http://192.0.2.1/',
             lang: 'ja',
         });
         expect(result).toEqual({
             success: false,
-            error: 'InternalServerError',
+            error: 'BadRequest',
         });
     }, 50000);
-    it('Bad request no dns ipv6 website test', async () => {
+    (isLocal ? it.skip : it)('Bad request no dns ipv6 website test', async () => {
         const result = await getMeta({
             url: 'http://[fe00::1]/',
             lang: 'ja',
         });
         expect(result).toEqual({
             success: false,
-            error: 'InternalServerError',
+            error: 'BadRequest',
         });
     }, 50000);
     it('Bad request invalid protocol test', async () => {
@@ -105,7 +108,7 @@ describe('getOgp Test', () => {
         });
         expect(result).toEqual({
             success: false,
-            error: 'InternalServerError',
+            error: 'BadRequest',
         });
     }, 50000);